Move telemetry drivers under observability
