+++ gestaltd/internal/daemon/factories.go
@@ -11,10 +11,10 @@ import (
 	"github.com/valon-technologies/gestalt/server/core"
 	"github.com/valon-technologies/gestalt/server/internal/bootstrap"
 	"github.com/valon-technologies/gestalt/server/internal/config"
-	telemetrynoop "github.com/valon-technologies/gestalt/server/internal/drivers/telemetry/noop"
-	telemetryotlp "github.com/valon-technologies/gestalt/server/internal/drivers/telemetry/otlp"
-	telemetrystdout "github.com/valon-technologies/gestalt/server/internal/drivers/telemetry/stdout"
 	"github.com/valon-technologies/gestalt/server/services/invocation"
+	telemetrynoop "github.com/valon-technologies/gestalt/server/services/observability/drivers/noop"
+	telemetryotlp "github.com/valon-technologies/gestalt/server/services/observability/drivers/otlp"
+	telemetrystdout "github.com/valon-technologies/gestalt/server/services/observability/drivers/stdout"
 	"github.com/valon-technologies/gestalt/server/services/operator"
 	"github.com/valon-technologies/gestalt/server/services/providerdrivers"
 	secretsenv "github.com/valon-technologies/gestalt/server/services/secrets/drivers/env"
