+++ gestaltd/services/observability/drivers/otlp/otlp.go
@@ -28,9 +28,9 @@ import (
 
 	"github.com/valon-technologies/gestalt/server/core"
 	"github.com/valon-technologies/gestalt/server/internal/bootstrap"
-	"github.com/valon-technologies/gestalt/server/internal/drivers/telemetry/metricspipeline"
-	telemetrystdout "github.com/valon-technologies/gestalt/server/internal/drivers/telemetry/stdout"
-	"github.com/valon-technologies/gestalt/server/internal/drivers/telemetry/telemetryutil"
+	"github.com/valon-technologies/gestalt/server/services/observability/drivers/metricspipeline"
+	telemetrystdout "github.com/valon-technologies/gestalt/server/services/observability/drivers/stdout"
+	"github.com/valon-technologies/gestalt/server/services/observability/drivers/telemetryutil"
 	"gopkg.in/yaml.v3"
 
 	"go.opentelemetry.io/contrib/bridges/otelslog"
