+++ gestaltd/services/observability/drivers/stdout/stdout_test.go
